+++ ConsoleTicTacToe/FancyGame.h
@@ -8,7 +8,7 @@ namespace tictactoe
 	class FancyGame : public GameSimulation
 	{
 	public:
-		FancyGame(uint32_t m, uint32_t n, uint32_t k);
+		FancyGame(uint16_t m, uint16_t n, uint16_t k);
 		virtual ~FancyGame();
 
 		virtual bool Update() override;
